# Avoid unnecessary double string split in ActiveSupport::MessageVerifier#verified