### Summary

<!-- Provide a general description of the code changes in your pull
request... were there any bugs you had fixed? If so, mention them. If
these bugs have open GitHub issues, be sure to tag them here as well,
to keep the conversation linked together. -->

`ActiveSupport::MessageVerifier#verified` is causing `signed_message` to be split twice: first inside `#valid_message?` and then inside `#verified`. This is ultimately unnecessary.

### Benchmark

I renamed the current implementation of `#valid_message?` and `#verified` to `#valid_message_old?` and `#verified_old` respectively, with the latter modified to call `#valid_message_old?`. I then ran the following benchmark:

```ruby
data = SecureRandom.alphanumeric(SIZE)
signed_message = ActiveSupport::MessageVerifier.new("Hey, I'm a secret!").generate(data)

GC.start

Benchmark.ips do |x|
  x.time = 10
  x.warmup = 2

  x.report('new verified') do
    @verifier.verified(signed_message)
  end
  x.report('old verified') do
    @verifier.verified_old(signed_message)
  end

  x.compare!
end

Benchmark.ips do |x|
  x.time = 10
  x.warmup = 2

  x.report('new valid message') do
    @verifier.valid_message?(signed_message)
  end
  x.report('old valid message') do
    @verifier.valid_message_old?(signed_message)
  end

  x.compare!
end
```

With `SIZE = 2**12` (4 kB, approximately the maximum size of a cookie):
```
Comparison:
        new verified:    28808.8 i/s
        old verified:    25400.1 i/s - 1.13x  (± 0.00) slower

Comparison:
        new verified:    37774.6 i/s
        old verified:    32482.2 i/s - 1.16x  (± 0.00) slower

Comparison:
        new verified:    38709.1 i/s
        old verified:    33416.9 i/s - 1.16x  (± 0.00) slower

Comparison:
        new verified:    38629.5 i/s
        old verified:    32879.7 i/s - 1.17x  (± 0.00) slower

Comparison:
   old valid message:    70262.7 i/s
   new valid message:    69663.3 i/s - same-ish: difference falls within error

Comparison:
   old valid message:    69522.6 i/s
   new valid message:    68408.8 i/s - same-ish: difference falls within error

Comparison:
   new valid message:    69642.7 i/s
   old valid message:    69283.1 i/s - same-ish: difference falls within error

Comparison:
   new valid message:    69779.8 i/s
   old valid message:    69510.5 i/s - same-ish: difference falls within error
```

With `SIZE = 2**22` (4 MB):
```
Comparison:
        new verified:       75.0 i/s
        old verified:       49.9 i/s - 1.50x  (± 0.00) slower

Comparison:
        new verified:       70.0 i/s
        old verified:       49.1 i/s - 1.43x  (± 0.00) slower

Comparison:
        new verified:       74.6 i/s
        old verified:       49.7 i/s - 1.50x  (± 0.00) slower

Comparison:
        new verified:       40.1 i/s
        old verified:       32.2 i/s - 1.25x  (± 0.00) slower

Comparison:
   new valid message:      133.3 i/s
   old valid message:      131.3 i/s - same-ish: difference falls within error

Comparison:
   old valid message:      131.5 i/s
   new valid message:      131.4 i/s - same-ish: difference falls within error

Comparison:
   new valid message:      133.3 i/s
   old valid message:      132.1 i/s - same-ish: difference falls within error

Comparison:
   new valid message:      132.6 i/s
   old valid message:      130.7 i/s - same-ish: difference falls within error
```

As expected, with a larger signed message, a greater speedup.

### Other Information

<!-- If there's anything else that's important and relevant to your pull
request, mention that information here. This could include
benchmarks, or other information.

If you are updating any of the CHANGELOG files or are asked to update the
CHANGELOG files by reviewers, please add the CHANGELOG entry at the top of the file.

Finally, if your pull request affects documentation or any non-code
changes, guidelines for those changes are [available
here](https://edgeguides.rubyonrails.org/contributing_to_ruby_on_rails.html#contributing-to-the-rails-documentation)

Thanks for contributing to Rails! -->

None.