@jcmfernandes 2021-08-31
You're absolutely right @p8! In fact, we can do better. We know what digest is being used, so we can know how many chars can it take to be represented in base64 (worst case). That means that we can use `#index` with an offset of something like `signed_message.length - maximum_digest_length - 2`. Makes sense? I'm going to investigate.