@jcmfernandes 2021-09-01
First, ignore the ugly commit history as I'll fix that before we merge.

What the benchmarks show is the much-increased resistance to invalid messages, those that don't feature the separator. The new code is not sensitive to the size of the input whenever it doesn't feature the separator.

Using `#rindex` to find the separator provides a similar performance when the separator is there. This makes sense, as it will have to cover at most 128 bytes when using SHA512. However, it's much slower when dealing with invalid messages that don't feature the separator.

So, we are showing a ~1.4x speedup for 4 kB valid messages and a ~1.8x speedup for 4 MB valid messages. When dealing with invalid messages, we are looking at ~3x to ~10x speedup for 4 kB messages and over 1000x speedup for 4 MB messages.

I dropped a few comments in the PR and would appreciate your feedback @p8. Thank you! :raised_hands: 

```
4 kB

-- OLD
Warming up --------------------------------------
           #verified     3.554k i/100ms
     #valid_message?     7.414k i/100ms
Calculating -------------------------------------
           #verified     36.783k (± 1.4%) i/s -    369.616k in  10.050797s
     #valid_message?     75.039k (± 1.7%) i/s -    756.228k in  10.080698s

-- NEW
Warming up --------------------------------------
           #verified     5.081k i/100ms
     #valid_message?    10.109k i/100ms
Calculating -------------------------------------
           #verified     53.090k (± 1.7%) i/s -    533.505k in  10.052230s
     #valid_message?     98.483k (± 1.1%) i/s -    990.682k in  10.060535s
```

```
4 MB

-- OLD
Warming up --------------------------------------
           #verified     6.000  i/100ms
     #valid_message?    15.000  i/100ms
Calculating -------------------------------------
           #verified     68.867  (± 2.9%) i/s -    690.000  in  10.030612s
     #valid_message?    158.845  (± 1.9%) i/s -      1.590k in  10.014114s

-- NEW
Warming up --------------------------------------
           #verified    12.000  i/100ms
     #valid_message?    27.000  i/100ms
Calculating -------------------------------------
           #verified    127.487  (± 2.4%) i/s -      1.284k in  10.079559s
     #valid_message?    271.107  (± 3.0%) i/s -      2.727k in  10.067583s
```

```
4 kB INVALID MESSAGE

-- OLD
Warming up --------------------------------------
           #verified    26.000k i/100ms
     #valid_message?    36.977k i/100ms
Calculating -------------------------------------
           #verified    259.917k (± 1.8%) i/s -      2.600M in  10.006742s
     #valid_message?    368.768k (± 0.4%) i/s -      3.698M in  10.027311s

-- NEW
Warming up --------------------------------------
           #verified    67.852k i/100ms
     #valid_message?   249.353k i/100ms
Calculating -------------------------------------
           #verified    676.537k (± 1.8%) i/s -      6.785M in  10.032982s
     #valid_message?      2.529M (± 2.1%) i/s -     25.434M in  10.061480s
```

```
4 MB INVALID MESSAGE

-- OLD
Warming up --------------------------------------
           #verified    48.000  i/100ms
     #valid_message?    48.000  i/100ms
Calculating -------------------------------------
           #verified    483.897  (± 0.8%) i/s -      4.848k in  10.019468s
     #valid_message?    481.756  (± 2.3%) i/s -      4.848k in  10.068306s

-- NEW
Warming up --------------------------------------
           #verified    66.592k i/100ms
     #valid_message?   255.351k i/100ms
Calculating -------------------------------------
           #verified    677.241k (± 2.2%) i/s -      6.792M in  10.034678s
     #valid_message?      2.503M (± 2.5%) i/s -     25.024M in  10.004185s
```